% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{add_font}
\alias{add_font}
\title{Add fonts}
\usage{
add_font(font)
}
\arguments{
\item{font}{a font file name}
}
\value{
A line of the GitHub Actions script to install the font package.
\code{NULL} if \code{font} is not recognized.
}
\description{
This function is called by \code{\link[=build_ghworkflow]{build_ghworkflow()}} to install fonts.
}
\keyword{internal}
