% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{add_hyphen}
\alias{add_hyphen}
\title{Add hyphenation patterns}
\usage{
add_hyphen(lang)
}
\arguments{
\item{lang}{a language code, such as "fr-FR"}
}
\value{
A line of the GitHub Actions script to install the hyphenation package.
\code{NULL} if \code{lang} is not recognized.
}
\description{
This function is called by \code{\link[=build_ghworkflow]{build_ghworkflow()}} to install hyphenation patterns.
}
\keyword{internal}
