% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{build_gitignore}
\alias{build_gitignore}
\title{Build .gitignore}
\usage{
build_gitignore()
}
\description{
Build a .gitignore file suitable for R Markdown projects.
}
\details{
The .gitignore file contains the list of files (file name patterns) that must not be controlled by git.
Run this function once in each project created from a memoiR template, before activating version control.
}
