% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{draft_memoir}
\alias{draft_memoir}
\title{RStudio Project}
\usage{
draft_memoir(path, ...)
}
\arguments{
\item{path}{the path to the newly created project}

\item{...}{extra arguments passed by the New Project Wizard}
}
\value{
Used for side effects.
}
\description{
This function is run by the RStudio project wizard to create a new document project.
}
\keyword{internal}
