% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effort.fitfeeder.R
\name{effort.fitfeeder}
\alias{effort.fitfeeder}
\title{Prepare National-scale Effort Data For MEMOIR.fit input
 
Takes data from a national-scale mark-recapture
database and formats it for input to MEMOIR.fit. Inputs are vectors from a table
equivalent to outputEffort (length P), which would be generated by MEMOIR.sim
for simulated populations.}
\usage{
effort.fitfeeder(YEAR, SPECIES, LOC, BANDED, splitBy)
}
\arguments{
\item{YEAR}{Numeric or date vector specifying the year in which animals were banded}

\item{SPECIES}{Text factor specifying the scientific name of the species}

\item{LOC}{Text factor specifying the location at which animals were banded}

\item{BANDED}{Integer. The number of animals banded in that year at that location}
}
\value{
The function returns a table formatted in the same way as MEMOIR.sim[[2]]
        c("recRepNo", "BAND", "SCIENA", "BANDED", "RECOVERED", "ELAPSED_DAYS", "BANDYEAR",
          "BANDLOC", "RECOVERYEAR"), for input to MEMOIR.fit as 'records'
}
\description{
Prepare National-scale Effort Data For MEMOIR.fit input
 
Takes data from a national-scale mark-recapture
database and formats it for input to MEMOIR.fit. Inputs are vectors from a table
equivalent to outputEffort (length P), which would be generated by MEMOIR.sim
for simulated populations.
}
\examples{
with(effort, effort.fitfeeder(YEAR, SPECIES, LOC, BANDED, splitBy))
}
