% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{Knit}
\alias{Knit}
\alias{knit_all}
\alias{knit_template}
\title{Knit}
\usage{
knit_all(destination = usethis::proj_path("docs"), gallery = "gallery")

knit_template(
  template,
  output_format,
  destination = usethis::proj_path("docs"),
  gallery = "gallery"
)
}
\arguments{
\item{destination}{name of the folder containing GitHub pages or equivalent.}

\item{gallery}{name of the subfolder of \code{destination} to store the knitted documents.}

\item{template}{name of the template to knit, e.g. "simple_article".}

\item{output_format}{TODO.}
}
\description{
Create documents from templates
}
\details{
These functions are used to test the templates and produce a gallery.
\itemize{
\item \code{knit_template()} produces an HTML and a PDF output of the chosen template.
\item \code{knit_all()} runs knit_template() on all templates of the package.
The \code{type} argument selects the way templates are rendered:
\item "document": HTML by \code{rmdformats::downcute} and PDF by \code{bookdown::pdf_book}.
\item "book": HTML by \code{bookdown::gitbook} and PDF by \code{bookdown::pdf_book}.
\item "slides": HTML by \code{rmarkdown::ioslides_presentation} and PDF by \code{rmarkdown::beamer_presentation}.
}
}
