% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\docType{package}
\name{memoiR-package}
\alias{memoiR}
\alias{memoiR-package}
\title{Package memoiR}
\description{
R Bookdown templates to publish documents,
especially relying on the memoir LaTeX package
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EricMarcon/memoiR}
  \item Report bugs at \url{https://github.com/EricMarcon/memoiR/issues}
}

}
\author{
\strong{Maintainer}: Eric Marcon \email{eric.marcon@agroparistech.fr} (\href{https://orcid.org/0000-0002-5249-321X}{ORCID})

}
