% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.boot.R
\name{MEMOIR.boot}
\alias{MEMOIR.boot}
\title{MEMOIR.boot(): Generate bootstrap estimates of annual adult survival, immature survival,
               first-year survival, and senescence for a single species.}
\usage{
MEMOIR.boot(
  records,
  effort,
  wearRate,
  breedingAge,
  zombieMode,
  deadOnly,
  niter = 1000
)
}
\arguments{
\item{records}{A data-frame of records, as for MEMOIR.fit()}

\item{effort}{A data-frame of annualised banding-effort, as for MEMOIR.fit()}

\item{wearRate}{A dataframe of wear-rate values, as for MEMOIR.fit()}

\item{niter}{A number of bootstrap iterations to run}
}
\description{
Returns a list with values [[1]]: The estimated adult survival, immature survival,
first-year survival and senescence for each iteration, and [[2]]: the SE of the
iterated survival and senescence estimates.
}
