% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.fit.R
\name{MEMOIR.fit}
\alias{MEMOIR.fit}
\title{Fit a (set of) MEMOIR model(s) from simulated or real-world data}
\usage{
MEMOIR.fit(records, effort, wearRates, groups, fitLocs = TRUE)
}
\arguments{
\item{records}{A dataframe of individual recapture records, formatted in the same
way as a outs$outputRecords frame from MEMOIR.sim.}

\item{effort}{A dataframe of annual (possibly location-specific) animal-marking effort,
formatted the same way as a outs$outputEffort frame from MEMOIR.sim.}

\item{wearRates}{A dataframe of species names and band-wear rates, used to calculate mark-
loss.}

\item{groups}{Factor. One level for each group which you wish to be modelled separately.
i.e., one level per species, if fitting species-specific estimates, or one
level per sex within each species, if sex-specific species-specific if
fitting separate curves for each sex within each species. If you are modelling
simulated data (from MEMOIR.sim()), groups will most likely be iteration
numbers, which can be extracted as unique(simName[[1]]$recRepNo).}
}
\value{
The function returns a dataframe of weights for each observation under various
        weighting criteria: Wa, Wi, Wb, Wal, Wil, Waib, and Waibl.
}
\description{
Estimates weightings for observations based on a prior model of the probability of 
making each observation. Weightings estimated are Wi (the odds of missing an
animal because of low research intensity, relative to the animal observed under
the most intense research), Wa (the odds of missing an animal at this animal's
age-at-observation, relative to the age-of-observation at which the greatest number
of marked animals could have been observed), and Wb (the odds of missing an observation
at this animal's age-at-observation owing to mark loss from wear, relative to the age
at which no marks would have been lost to wear). Location-specific forms of Wa and Wi
(Wal and Wil, respectively), and combined-odds weights (Waib and Wbail) are
also calculated. Combined-odds weights are applied to points, and these are
used to estimate each species' mortality curve.
Note that no fit will be attempted on for species with fewer than 20 total recoveries.
This behaviour is hard-set at line 92 (in version 0.1).
Known (harmless) bug: You will likely see the error message:
In max(yset$n.risk) : no non-missing arguments to max; returning -Inf
This occurs when there is an age-class in which no individuals are recorded dead.
-Inf values are automatically detected and corrected.
}
\examples{
simName <- MEMOIR.sim()
fitName <- MEMOIR.fit(records=simName[[1]], effort=simName[[2]], wearRates=simName[[3]], groups=unique(simName[[1]]$recRepNo), fitLocs=TRUE)
}
