% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.rmse.R
\name{MEMOIR.rmse}
\alias{MEMOIR.rmse}
\title{MEMOIR.rmse(): calculate root mean squared error of a series of survival estimates
               generated by MEMOIR.simplotter() at a set of given sampling ages with
               a set of given true values.}
\usage{
MEMOIR.rmse(estimates, ages = c(5, 10, 15), proportions = c(0.75, 0.5, 0.25))
}
\arguments{
\item{estimates}{A dataframe of fitted model predictions from MEMOIR.simplotter(). Must
have names as c("iter", "ages", "numberAlive", "propAlive"), where
iter is iteration number, ages are ages (in days), numberAlive is the
fitted-model number of animals alive, and propAlive is the fitted-model
proportion of animals alive.}

\item{ages}{A vector of ages (in years) at which to estimate the root mean squared error
of the estimate. Defaults to c(5,10,15).}

\item{proportions}{A vector of (true) proportional survival at the ages given in ages.
Defaults to c(0.75, 0.5, 0.25), which are the survival rates for
Curve A (Baylis, PhD thesis) at 5, 10, and 15 years of age.}
}
\description{
Returns the Root Mean Squared Error of the estimate at the given sampling ages. Output is
formatted as a dataframe, with ages as column one, proportional survival as column two,
and RMSE as column three.
}
