% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.senesc.R
\name{MEMOIR.senesc}
\alias{MEMOIR.senesc}
\title{MEMOIR.senesc: estimate species' senescence rates from a MEMOIR.survrates() fit.}
\usage{
MEMOIR.senesc(survrates, startAge = 2)
}
\arguments{
\item{survrates}{frame [[1]] from a MEMOIR.survrates output. If this has glitched and
produced all NAs for 'species', it is necessary to edit the species
names in.}

\item{startAges}{a vector of ages at which senescence could start. Generally equal
                 to the age at first breeding, for each species.
Currently outputs a dataframe with species name (as in the 'survrates' input),
'coef' (the estimated senescence parameter), and 'SE' (the SE of the senescence parameter).}
}
\description{
MEMOIR.senesc: estimate species' senescence rates from a MEMOIR.survrates() fit.
}

