% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.sim.R
\name{MEMOIR.sim}
\alias{MEMOIR.sim}
\title{Simulate National-Scale Mark-Recapture}
\usage{
MEMOIR.sim(
  locs = 5,
  mortCurve = c("A"),
  history = 64,
  batchProb = 0.4,
  catchSize = 50,
  researchProp = 0.5,
  wearRate = 2.22,
  maxWear = 65,
  censorProb = 0.4,
  liveRecapDeflator = 0.5,
  niter = 1000
)
}
\arguments{
\item{locs}{Integer. Number of locations ('locodes', or equivalent) to simulate}

\item{history}{Integer. Number of years between first (potential) band-application and
the end of data-collection.}

\item{batchProb}{Numeric. 0 < batchProb < Inf. Denotes the mean number of research
projects initiated for the species in a given year. Number of
studies initiated per year is a random draw from a Poisson distribution,
batches ~ rpois(1, batchProb)}

\item{catchSize}{Integer. Max number of animals caught (banded) in a single research
trip. Actual numbers caught on a trip are sampled from a Uniform dist,
runif(1, 1, catchSize)}

\item{researchProp}{Numeric. Generally between zero and 1, but will accept values higher
than one. Higher values increase the proportion of dead recaptures
are research-related, and therefore increases the degree to which
recaupture rate within animals from a location correlates with
research effort at that location}

\item{wearRate}{Numeric. The mean rate at which bands lose mass to wear, expressed as a
percentage of their starting weight per year.}

\item{maxWear}{Numeric. The percent mass-loss at which bands are lost to wear.}

\item{censorProb}{Numeric. The probability that an individual observation is 'censored',
i.e., from an individual whose last resighting was a live-resighting. Reasonable
values can be obtained as in Baylis et al (2014), as described for 'Proportion
of live recaptures'.}

\item{liveRecapDeflator}{Numeric. Adjusts the probability of resighting a live,
marked animal relative to the probability of observing that animal if it had
died in that year. Can range from 0 to Inf, where 0 is 'there are no live
resightings', and Inf is 'every individual will be resighted alive in every year
that it is alive'. Live recaptures are modelled as arising only from research
activity, as would be the case for species marked with only metal bands,
requiring recapture for identification.}

\item{niter}{Numeric. The number of simulation iterations to be run. If the settings
result in some simulations where no animals are marked, those
iterations will go unrecorded in the output. After a simulation,
[niter] - length(unique([sim name][[1]]$recRepNo)) returns the
number of iterations in which no animals were marked.}

\item{line}{Single factor. Denotes the 'true' death curve to be used in simulations.
One of c("A", "B", "C", "D", "E"). See Baylis et al (2014) for curves.}
}
\value{
The function returns a list containing two data frames.
   1: Records. Details for every mark that was recovered in the simulation
   1a: recRepNo: shows which iteration the band was from
   1b: BAND: A unique seven-digit integer for each banded animal
   1c: SCIENA: The 'scientific name' of the animal. Placeholder value for read-in
               to MEMOIR.fit or similar functions: 'Animalis artificialis'
   1d: BANDED: The date the animal was banded, as "%Y-%m-%d"
   1e: RECOVERED: The date the animal was recovered, as "%Y-%m-%d"
   1f: ELAPSED_DAYS: The number of days between BANDED and RECOVERED
   1g: BANDYEAR: The year the animal was banded, as "%Y"
   1h: BANDLOC: The location the animal was banded in, as numeric. Used as a 'locode'
   1i: RECOVERYEAR: The year the animal was recovered, as "%Y"
   2: Effort. Details of when and where animal-marking effort took place
   2a: eRepNo: Shows which iteration the rest of the line's data was from
   2b: YEAR: The year in which bands were applied, as "%Y"
   2c: CAVS: Depreciated. A vector of Poisson-distributed random numbers
   2d: SPECIES: The 'scientific name' of the animals. Placeholder for read-in to
                 MEMOIR.fit or similar functions: 'Animalis artificialis'
   2e: BANDED: The total number of animals banded in YEAR
   2f: X1 - Xn: The number of animals banded at each of n simulated locodes in YEAR
}
\description{
Simulates data collected by national-scale mark-recovery schemes
for a species with multiple populations with limited inter-population movement.
Includes support for a variable number of populations, setting 'true' death-curve,
setting the number of years the species has been (potentially) monitored for at
each location, setting the distribution of number of individuals marked each time
a research team goes to mark animals, setting the relative likelihood of recapture
by researchers, or recapture by casual observers, setting the species' band-wear 
rate, and setting the amount of wear which causes mark-loss.
}
\examples{
MEMOIR.sim(locs=5, mortCurve=c("A"), history=64, batchProb=0.3, catchSize=50,
           researchProp=0.5, wearRate=0.22, maxWear=65, niter=1000)
}
