% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.simplotterV2.R
\name{MEMOIR.simplotterV2}
\alias{MEMOIR.simplotterV2}
\title{Generate plots and densely-sampled estimated survival curves from a fitted MEMOIR model
with observations from multiple locations or populations}
\usage{
MEMOIR.simplotterV2(x, zombieMode = TRUE, deadOnly = FALSE,
  PMType = "weibull", displayType = "kaplan")
}
\arguments{
\item{x}{A dataframe generated using MEMOIR.fit()[[1]]}

\item{zombieMode}{Convenience function, re-codes CENS values by flipping all zeroes to
ones and vice-versa. Implemented because MEMOIR.sim() codes live and
dead recaptures using opposite coding to the functions in package
Survival by default. TRUE by default.}

\item{deadOnly}{When TRUE, excludes live records from the dataset before analysis. Can be
useful to prevent bias arising from informative censoring.}

\item{PMType}{Type of parametric model to use in location-specific sub-models. Must be
a distribution included in survreg() from package Survival, so currently must
be one of "weibull", "exponential", "gaussian", "logistic", "lognormal",
or "loglogistic".}

\item{displayType}{Type of fit to plot. One of "kaplan" or "parametric". If "parametric",
the plot will display models based on whatever parametric distribution
was chosen in PMType.}
}
\value{
The function returns a list with two dataframes. The first dataframe has estimated
        survival by age for each iteration based on non-parametric Kaplan-Meier estimates.
        The second dataframe has estimated survival by age for each iteration based on
        whatever parametric distribution was selected in PMType. Also returns a plot, as
        chosen in displayType.
}
\description{
Estimates survival curves within each population using both non-parametric Kaplan-Meier
estimation and a parametric model of choice. Returns a figure showing the estimated
mortality curve for each iteration in simulated data, or a single estimated mortality
curve in real species data. For discussion of the different models implemented here, see
Baylis et al (xxxx) and Baylis S. M. (2016) [PhD title goes here].
}

