% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEMOIR.survrates.R
\name{MEMOIR.survrates}
\alias{MEMOIR.survrates}
\title{Estimate survival rates through each year-class for a set of species, using curves
from MEMOIR.simplotter()}
\usage{
MEMOIR.survrates(x, species, breedingAges)
}
\arguments{
\item{x}{Output 'KMOuts' from MEMOIR.simplotter() or output 'outs' from MEMOIR.simplotter.}

\item{species}{A vector of species names to estimate survival rates for separately.}

\item{breedingAges}{A vector of ages at first breeding, of the same length as parameter
species. These are taken as the break-points between 'juvenile' and
'adult' survival. So, for a species with breeding age listed as 2,
'juvenile' survival will be calculated as n-weighted mean survival
across age-ranges 0 - 1 years and 1 - 2 years, and 'adult' survival
will be calculated as n-weighted mean survival across all age-ranges
2 - 3 years or greater.}
}
\description{
Estimates juvenile and adult survival rates for each age-class, for each of a set of
species, using output from MEMOIR.simplotter(). Outputs a list with estimated survival
rates through each age-class for each species (as [[1]]) and N-weighted survival estimates
through all juvenile age-classes, and all adult age-classes, for each species (as [[2]]).
N-weighting is achieved by multiplying estimated survival through each age-class by the
(weighted) number of observations from that age-class, and dividing the sum of those
products by the sum of the (weighted) number of observations across all age-classes.
}

