% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.fitfeeder.R
\name{records.fitfeeder}
\alias{records.fitfeeder}
\title{Prepare National-Scale Records Data For MEMOIR.fit Input}
\usage{
records.fitfeeder(
  BAND,
  SCIENA,
  BANDED,
  RECOVERED,
  BANDLOC,
  SplitBy = c("SCIENA")
)
}
\arguments{
\item{BAND}{Numeric or text vector of length N. Unique codes identifying each individual.}

\item{SCIENA}{Text vector of length N. Binomial scientific name identifying the individual's
species.}

\item{BANDED}{Date vector of length N, formatted as YYYY-mm-dd. The date the animal was
banded.}

\item{RECOVERED}{Date vector of length N, formatted as YYYY-mm-dd. The date the animal
was (last) recovered or observed.}

\item{BANDLOC}{Text or numeric vector of length N. Unique values for each marking location
(i.e., Australian 'locode', or equivalent). Values must be coded the same
way as in effort.fitfeeder for input to MEMOIR.fit.}

\item{SplitBy}{One of c("BANDLOC", "SCIENA"). Indicates which groups will be used for
separate models in MEMOIR-fit. "SCIENA" gives separate curves for each
species, and "BANDLOC" give separate curves for each location.}
}
\value{
The function returns a table formatted in the same way as MEMOIR.sim[[1]]
        c("recRepNo", "BAND", "SCIENA", "BANDED", "RECOVERED", "ELAPSED_DAYS", "BANDYEAR",
          "BANDLOC", "RECOVERYEAR"), for input to MEMOIR.fit as 'records'.
}
\description{
Takes data from a national-scale mark-recapture
database and formats it for input to MEMOIR.fit. Inputs are vectors from a table
equivalent to outputRecords (length N), which would be generated by MEMOIR.sim
for simulated populations.
}
\examples{
with(records, records.fitfeeder(BAND=bandNo, SCIENA=sciName, BANDED=bandDate,
     RECOVERED=recoverDate, BANDLOC=locode)) 
}
