% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wear.fitfeeder.R
\name{wear.fitfeeder}
\alias{wear.fitfeeder}
\title{Prepare a band-wear table for MEMOIR.fit read-in}
\usage{
wear.fitfeeder(SPECIES, rate)
}
\arguments{
\item{SPECIES}{A factor containing each species name for which fitted estimates are needed}

\item{rate}{Either a vector with the same length as SPECIES, or a single number,
representing the wear rates for each species, or the shared wear-rate across
all species.}
}
\value{
The function returns a table formatted in the same way as MEMOIR.sim[[3]]
        c("SPECIES", "usedRate"), for input to MEMOIR.fit as 'records'.
}
\description{
Takes a vector of species names and a single wear-rate or vector of wear-rates, and 
outputs it for input to MEMOIR.fit.
}
\examples{
with(records, records.fitfeeder(BAND=bandNo, SCIENA=sciName, BANDED=bandDate,
     RECOVERED=recoverDate, BANDLOC=locode)) 
}
