% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{build_ghworkflow}
\alias{build_ghworkflow}
\title{Build GitHub Action Workflow}
\usage{
build_ghworkflow()
}
\description{
Build a YAML file (\code{.gihub/workflows/memoir.yml}) to knit the documents
of the project to GitHub Pages.
The workflow knits all R Markdown files according their header: all output
formats are produced and stored into the \code{gh-pages} branch of the project.
}
\details{
All HTML outputs have the same name so the last one knitted overwrites the
previous ones.
Keep only one HTML format in the header of each RMarkdown file.

No \code{DESCRIPTION} file is necessary in the project to install packages.
They must be declared in the options code chunk of each .Rmd file
(index.Rmd for the memoir template).

Two secrets must have been stored in the GitHub account:
\itemize{
\item GH_PAT: a valid access token,
\item EMAIL: the email address to send the workflow results to.
}
}
