% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{build_githubpages}
\alias{build_githubpages}
\title{Build GitHub Pages}
\usage{
build_githubpages(destination = usethis::proj_path("docs"))
}
\arguments{
\item{destination}{destination folder of the knitted documents.}
}
\description{
Copy the files produced by knitting to the destination folder.
}
\details{
Produced files are HTML pages and their companions (css, figures, libraries) and PDF documents.
}
