% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{build_githubpages}
\alias{build_githubpages}
\title{Build GitHub Pages}
\usage{
build_githubpages(destination = usethis::proj_path("docs"))
}
\arguments{
\item{destination}{destination folder of the knitted documents.}
}
\description{
Copy the files produced by knitting to the destination folder.
}
\details{
Produced files are HTML pages and their companions (css, figures, libraries) and PDF documents.
The function moves them all and the \code{README.md} file into the destination folder.
GitHub Pages allow making a website to present them:
\itemize{
\item \code{README.md} is the home page. Make it with \code{\link[=build_readme]{build_readme()}} to have links to the HTML and PDF outputs.
\item knit both HTML and PDF versions to avoid dead links.
\item run \code{build_githubpages()} when a document is knitted to move the outputs into the \code{docs} folder.
\item push to GitHub and activate GitHub Pages on the main branch and the \code{docs} folder.
The function is useless in book projects: the \emph{Build the Book} (i.e. the \code{\link[bookdown:render_book]{bookdown::render_book()}} function) takes care of every step.
}
}
\examples{
# Save working directory
original_wd <- getwd()
# Get a temporary working directory
wd <- tempfile("example")
# Simulate File > New File > R Markdown... > From Template > Stylish Article
rmarkdown::draft(wd, template="stylish_article", package="memoiR", edit=FALSE)
# Go to temp directory
setwd(wd)
# Make it the current project
usethis::proj_set(path = ".", force = TRUE)
# Build .gitignore
build_gitignore()
# render: knit to downcute
rmarkdown::render(input=list.files(pattern="*.Rmd"), 
                  output_format="rmdformats::downcute")
# Build index, HTML only
build_readme(PDF=FALSE)
# Build GitHub Pages
build_githubpages(destination="docs")
# List the GitHub Pages files
setwd("docs")
list.files(recursive=TRUE)
# Return to the original working directory and clean up
setwd(original_wd)
unlink(wd, recursive=TRUE)
}
