% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\name{Knit}
\alias{Knit}
\alias{knit_all}
\alias{knit_template}
\title{Knit}
\usage{
knit_all(destination = usethis::proj_path("docs"), gallery = "gallery")

knit_template(
  template,
  type,
  destination = usethis::proj_path("docs"),
  gallery = "gallery"
)
}
\arguments{
\item{destination}{name of the folder containing GitHub pages or equivalent.}

\item{gallery}{name of the subfolder of destination to store the knitted documents.}

\item{template}{name of the template to knit.}

\item{type}{type of the template to knit. May be "document", "book" or "slides".}
}
\description{
Create documents from templates
}
\details{
Used to test the templates
}
