% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memoiR.R
\docType{package}
\name{memoiR-package}
\alias{memoiR}
\alias{memoiR-package}
\title{memoiR: R Markdown and Bookdown Templates to Publish Documents}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Producing high-quality documents suitable for publication directly from R is made possible by the R Markdown ecosystem. 'memoiR' makes it easy. It provides templates to knit memoirs, articles and slideshows with helpers to publish the documents on GitHub Pages and activate continuous integration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ericmarcon.github.io/memoiR/}
  \item \url{https://github.com/EricMarcon/memoiR}
  \item Report bugs at \url{https://github.com/EricMarcon/memoiR/issues}
}

}
\author{
\strong{Maintainer}: Eric Marcon \email{eric.marcon@agroparistech.fr} (\href{https://orcid.org/0000-0002-5249-321X}{ORCID})

}
\keyword{internal}
